fix: move PromptLayout to the common folder, add icons
