+++ x-pack/platform/plugins/shared/agent_builder/public/application/components/access/prompts/upgrade_license_prompt.tsx
@@ -8,7 +8,7 @@
 import { EuiButton, EuiButtonEmpty, useEuiTheme } from '@elastic/eui';
 import React from 'react';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { PromptLayout, type PromptLayoutVariant } from './prompt_layout';
+import { PromptLayout, type PromptLayoutVariant } from '../../common/prompt_layout';
 import { useAgentBuilderServices } from '../../../hooks/use_agent_builder_service';
 import { useAssetBasePath } from '../../../hooks/use_asset_base_path';
